refactor(epaper): extract SD blob cache into shared module
Move the reTerminal E1003 driver's microSD blob-cache implementation
(mount/unmount, redirect resolve, image-id parse, read/write, clear, and
the pending-buffer write-back lifecycle) out of the panel driver and into
a dedicated device_classes/epaper/epaper_sd_cache.{h,cpp} module.

The driver now owns only board-specific concerns: it captures the SPI
bus, CS/EN/DET pins, and a restore_panel_bus callback via
epaper_sd_cache_init() in begin(), and forwards the SD-cache HAL surface
to the module. The module is gated on EPAPER_SD_CS_PIN and aggregated
into the build through epaper_device_class.cpp.

On boards without a shared-bus SD slot the three SD-cache HAL functions
resolve to inline no-ops provided by epaper_sd_cache.h (pulled in by
epaper_driver.h), so per-driver stubs are no longer needed -- the
Inkplate 5V2 stubs are removed. No functional change; blob cache key and
on-card /cache/<id>.g16p layout are unchanged.
