+++ src/app/device_classes/epaper/drivers/inkplate5v2_driver.cpp
@@ -99,10 +99,8 @@ uint16_t epaper_driver_battery_mv() {
 		return (uint16_t)(v * 1000.0);
 }
 
-// SD image cache is unsupported on this board (no shared-bus SD slot) — stubs.
-void epaper_driver_set_sd_cache_enabled(bool /*enabled*/) {}
-void epaper_driver_cache_flush() {}
-bool epaper_driver_sd_cache_clear() { return false; }
+// SD image cache is unsupported on this board (no shared-bus SD slot); the
+// SD-cache HAL vtable resolves to the inline no-ops in epaper_sd_cache.h.
 
 // ---------------------------------------------------------------------------
 // GFX primitives — pass-through to Adafruit_GFX methods inherited by Inkplate.
